Updated tests to work with new API
Note: The test 'Check PeriodicSyncManager API periodicSync.minPossiblePeriod exists as a number" is flaky.
It generally fails on the initial run, but succeeds on every subsequent run.

It is also not possible to test a lot of the one shot sync functionality programmatically since it depends
on losing/regaining a network connection which cannot be simulated in the javascript test framework.